add p2p config

+++ internal/p2p/options.go
@@ -31,7 +31,7 @@ func MultiAddressBuilder(ipAddr string, port uint) (ma.Multiaddr, error) {
 // buildOptions for the libp2p host.
 func (s *Service) buildOptions(ip net.IP, priKey *ecdsa.PrivateKey) []libp2p.Option {
 	cfg := s.cfg
-	listen, err := MultiAddressBuilder(ip.String(), cfg.TCPPort)
+	listen, err := MultiAddressBuilder(ip.String(), uint(cfg.TCPPort))
 	if err != nil {
 		log.Crit("Failed to p2p listen", "err", err)
 	}
@@ -39,7 +39,7 @@ func (s *Service) buildOptions(ip net.IP, priKey *ecdsa.PrivateKey) []libp2p.Opt
 		if net.ParseIP(cfg.LocalIP) == nil {
 			log.Crit(fmt.Sprintf("Invalid local ip provided: %s", cfg.LocalIP))
 		}
-		listen, err = MultiAddressBuilder(cfg.LocalIP, cfg.TCPPort)
+		listen, err = MultiAddressBuilder(cfg.LocalIP, uint(cfg.TCPPort))
 		if err != nil {
 			log.Crit("Failed to p2p listen", "err", err)
 		}
@@ -77,7 +77,7 @@ func (s *Service) buildOptions(ip net.IP, priKey *ecdsa.PrivateKey) []libp2p.Opt
 	}
 	if cfg.HostAddress != "" {
 		options = append(options, libp2p.AddrsFactory(func(addrs []ma.Multiaddr) []ma.Multiaddr {
-			external, err := MultiAddressBuilder(cfg.HostAddress, cfg.TCPPort)
+			external, err := MultiAddressBuilder(cfg.HostAddress, uint(cfg.TCPPort))
 			if err != nil {
 				log.Error("Unable to create external multiaddress", "err", err)
 			} else {
